requirements.txt
streamlit
faiss-cpu
PyMuPDF
sentence-transformers
transformers
torch